requirements.txt
Flask
Pillow
python-magic



